refactor : couple 도메인 서비스의 외부 접근을 분리
